Rename : home 내 component rename
