+++ client/package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "axios": "^1.2.2",
     "react": "^18.2.0",
+    "react-cookie": "^4.1.1",
     "react-dom": "^18.2.0",
     "react-icons": "^4.7.1",
     "react-router-dom": "^6.6.1",
