+++ client/src/components/Home/FeedMenu.js
@@ -2,8 +2,8 @@ import styled from "styled-components";
 
 const Wrapper = styled.div`
     position: absolute;
-    top: 83px;
-    right: 0;
+    top: 75px;
+    right: 1px;
     display: flex;
     flex-direction: column;
     width: 100px;
@@ -15,7 +15,7 @@ const Wrapper = styled.div`
     button {
         height: 100%;
         background-color: white;
-        opacity: 0.4;
+        opacity: 0.9;
         border: 0;
         border-bottom: 1px solid #dbdbdb;
         font-size: 14px;
