+++ client/src/components/public/Post/EditPost.js
@@ -1,0 +1,7 @@
+const EditPost = () => {
+    return (
+        <></>
+    )
+};
+
+export default EditPost;
